+++ src/zrb/builtin/__init__.py
@@ -1,4 +1,4 @@
-from zrb.builtin.base64 import decode_base64, encode_base64
+from zrb.builtin.base64 import decode_base64, encode_base64, validate_base64
 from zrb.builtin.git import (
     get_git_diff,
     git_commit,
@@ -7,8 +7,10 @@ from zrb.builtin.git import (
     prune_local_branches,
 )
 from zrb.builtin.git_subtree import git_add_subtree, git_pull_subtree, git_push_subtree
+from zrb.builtin.http import generate_curl, http_request
+from zrb.builtin.jwt import decode_jwt, encode_jwt, validate_jwt
 from zrb.builtin.llm.llm_chat import llm_chat
-from zrb.builtin.md5 import hash_md5, sum_md5
+from zrb.builtin.md5 import hash_md5, sum_md5, validate_md5
 from zrb.builtin.project.add.fastapp.fastapp_task import add_fastapp_to_project
 from zrb.builtin.project.create.project_task import create_project
 from zrb.builtin.python import format_python_code
@@ -30,6 +32,17 @@ from zrb.builtin.todo import (
     log_todo,
     show_todo,
 )
+from zrb.builtin.uuid import (
+    generate_uuid_v1,
+    generate_uuid_v3,
+    generate_uuid_v4,
+    generate_uuid_v5,
+    validate_uuid,
+    validate_uuid_v1,
+    validate_uuid_v3,
+    validate_uuid_v4,
+    validate_uuid_v5,
+)
 
 assert create_project
 assert add_fastapp_to_project
@@ -38,9 +51,14 @@ assert make_bash_autocomplete
 assert make_zsh_autocomplete
 assert encode_base64
 assert decode_base64
+assert validate_base64
+assert encode_jwt
+assert decode_jwt
+assert validate_jwt
 assert llm_chat
 assert hash_md5
 assert sum_md5
+assert validate_md5
 assert get_git_diff
 assert prune_local_branches
 assert format_python_code
@@ -64,3 +82,14 @@ assert setup_latex_on_ubuntu
 assert setup_asdf
 assert setup_tmux
 assert setup_zsh
+assert validate_uuid
+assert validate_uuid_v1
+assert validate_uuid_v3
+assert validate_uuid_v4
+assert validate_uuid_v5
+assert generate_uuid_v1
+assert generate_uuid_v3
+assert generate_uuid_v4
+assert generate_uuid_v5
+assert http_request
+assert generate_curl